requirements.txt
llama-cpp-python==0.2.24
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.2.2
huggingface_hub==0.14.1